Move SearchBar to own file and fix test
